NSInteger transition
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2268 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/DataBasin/DBObjectInspector.m
@@ -150,12 +150,12 @@
 /** --- Data Source --- **/
 
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
   return [arrayRows count];
 }
 
-- (id) tableView: (NSTableView*)aTableView objectValueForTableColumn: (NSTableColumn*)column row: (int)rowIndex
+- (id) tableView: (NSTableView*)aTableView objectValueForTableColumn: (NSTableColumn*)column row: (NSInteger)rowIndex
 {
   id retObj;
   NSDictionary *row;
